+++ client/src/router.js
@@ -6,6 +6,7 @@ import Login from './views/Login.vue'
 import ErrorVue from './views/Error.vue'
 import Schema from './views/Schema.vue'
 import Groups from './views/Groups.vue'
+import Repos from './views/Repos.vue'
 import Users from './views/Users.vue'
 import Ldap from './views/Ldap.vue'
 import Awx from './views/Awx.vue'
@@ -13,6 +14,7 @@ import Designer from './views/Designer.vue'
 import Credentials from './views/Credentials.vue'
 import Sshkey from './views/Sshkey.vue'
 import Jobs from './views/Jobs.vue'
+import Logs from './views/Logs.vue'
 import Profile from './views/Profile.vue'
 import TokenStorage from './lib/TokenStorage.js'
 Vue.use(Router);
@@ -65,6 +67,12 @@ export default new Router({
         beforeEnter: checkAdmin
       },
       {
+        path:"/repos",
+        name:"Repos",
+        component:Repos,
+        beforeEnter: checkAdmin
+      },
+      {
         path:"/users",
         name:"Users",
         component:Users,
@@ -92,6 +100,12 @@ export default new Router({
         path:"/sshkey",
         name:"Sshkey",
         component:Sshkey,
+        beforeEnter: checkAdmin
+      },
+      {
+        path:"/logs",
+        name:"Logs",
+        component:Logs,
         beforeEnter: checkAdmin
       },
       {