+++ client/src/views/Jobs.vue
@@ -14,32 +14,53 @@
         <table v-if="isLoaded" class="table is-bordered is-narrow">
           <thead>
             <tr class="has-text-left">
-              <th style="width:100px" v-if="isAdmin">Action</th>
-              <th style="width:100px">id</th>
+              <th class="action" v-if="isAdmin"></th>
+              <th class="id">id</th>
               <th>form</th>
-              <th>job type</th>
+              <th class="jobtype">job type</th>
               <th>target</th>
-              <th>status</th>
+              <th class="status">status</th>
               <th>start time</th>
               <th>end time</th>
               <th>user</th>
             </tr>
           </thead>
           <tbody>
-          <tr v-for="j in displayedJobs" :key="j.id" :class="{'has-background-success-light':(j.status=='success' && j.id!=jobId),'has-background-danger-light':(j.status=='failed' && j.id!=jobId),'has-background-warning':(j.status=='aborted' && j.id!=jobId),'has-background-info':j.id==jobId,'has-text-white':j.id==jobId}">
-            <td v-if="isAdmin" class="has-background-info-light"><span class="icon has-text-danger is-clickable" @click="tempJobId=j.id;showDelete=true" title="Delete job"><font-awesome-icon icon="trash-alt" /></span></td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.id}}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.form}}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.job_type || "ansible" }}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.target}}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.status}}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.start | moment('YYYY-MM-DD HH:mm:ss')}}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.end | moment('YYYY-MM-DD HH:mm:ss')}}</td>
-            <td class="is-clickable has-text-left" @click="loadOutput(j.id)">{{j.user}} ({{j.user_type}})</td>
-          </tr>
+          <template v-for="j in displayedJobs">
+            <tr :key="j.id" :class="{'has-background-success-light':(j.status=='success' && j.id!=jobId),'has-background-danger-light':(j.status=='failed' && j.id!=jobId),'has-background-warning-light':((j.status=='aborted'||j.status=='warning') && j.id!=jobId),'has-background-info':j.id==jobId,'has-text-white':j.id==jobId}">
+              <td v-if="isAdmin" class="has-background-info-light"><span class="icon has-text-danger is-clickable" @click="tempJobId=j.id;showDelete=true" title="Delete job"><font-awesome-icon icon="trash-alt" /></span></td>
+              <td class="is-clickable has-text-left" @click="toggleCollapse(j.id)">
+                <span>{{j.id}}</span>
+                <template v-if="j.job_type=='multistep'">
+                  <span class="icon is-pulled-right" v-if="!collapsed.includes(j.id)"><font-awesome-icon icon="caret-right" /></span>
+                  <span class="icon is-pulled-right" v-else><font-awesome-icon icon="caret-down" /></span>
+                </template>
+              </td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.form">{{j.form}}</td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.job_type">{{j.job_type || "ansible" }}</td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.target">{{j.target}}</td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.status">{{j.status}}</td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.start">{{j.start | moment('YYYY-MM-DD HH:mm:ss')}}</td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.end">{{j.end | moment('YYYY-MM-DD HH:mm:ss')}}</td>
+              <td class="is-clickable has-text-left" @click="loadOutput(j.id)" :title="j.user">{{j.user}} ({{j.user_type}})</td>
+            </tr>
+            <template v-for="c in childJobs(j.id)">
+              <tr :key="c.id" :class="{'has-background-success-light':(c.status=='success' && c.id!=jobId),'has-background-danger-light':(c.status=='failed' && c.id!=jobId),'has-background-warning-light':((c.status=='aborted'||c.status=='warning') && c.id!=jobId),'has-background-info':c.id==jobId,'has-text-white':c.id==jobId}">
+                <td v-if="isAdmin" class="has-background-info-light"><span class="icon has-text-danger is-clickable" @click="tempJobId=c.id;showDelete=true" title="Delete job"><font-awesome-icon icon="trash-alt" /></span></td>
+                <td class="is-clickable has-text-right" @click="loadOutput(c.id)">{{c.id}}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.form">{{c.form}}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.job_type">{{c.job_type || "ansible" }}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.target">{{c.target}}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.status">{{c.status}}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.start">{{c.start | moment('YYYY-MM-DD HH:mm:ss')}}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.end">{{c.end | moment('YYYY-MM-DD HH:mm:ss')}}</td>
+                <td class="is-clickable has-text-left" @click="loadOutput(c.id)" :title="c.user">{{c.user}} ({{c.user_type}})</td>
+              </tr>
+            </template>
+          </template>
           </tbody>
         </table>
-        <nav v-if="isLoaded && jobs.length>perPage" class="pagination" role="pagination" aria-label="pagination">
+        <nav v-if="isLoaded && parentJobs.length>perPage" class="pagination" role="pagination" aria-label="pagination">
           <a class="pagination-previous" v-if="page != 1" @click="page--">Previous</a>
           <a class="pagination-next" @click="page++" v-if="page < pages.length">Next page</a>
           <ul class="pagination-list">
@@ -137,7 +158,8 @@
         jobId:undefined,
         showExtraVars:false,
         tempJobId:undefined,
-        showDelete:false
+        showDelete:false,
+        collapsed:[]
       }
     },
     methods:{
@@ -149,6 +171,22 @@
         }catch(e){
           this.$toast.error("Error copying to clipboard : \n" + e)
         }
+      },
+      toggleCollapse(id){
+        if(!this.collapsed.includes(id)){
+          this.collapsed.push(id)
+        }else{
+          for( var i = 0; i < this.collapsed.length; i++){
+              if ( this.collapsed[i] === id) {
+                  this.collapsed.splice(i, 1);
+                  i--;
+              }
+          }
+        }
+      },
+      childJobs(id){
+        var ref=this
+        return this.jobs.filter(x=> (x.parent_id===id && ref.collapsed.includes(id))).sort((a, b) => a.id > b.id && 1 || -1)
       },
       loadJobs(){
         var ref= this;
@@ -168,7 +206,7 @@
       },
       setPages () {
         this.pages=[]
-        let numberOfPages = Math.ceil(this.jobs.length / this.perPage);
+        let numberOfPages = Math.ceil(this.parentJobs.length / this.perPage);
         for (let index = 1; index <= numberOfPages; index++) {
          this.pages.push(index);
         }
@@ -220,8 +258,11 @@
       }
     },
     computed: {
+      parentJobs (){
+        return this.jobs.filter(x => !x.parent_id)
+      },
       displayedJobs () {
-        return this.paginate(this.jobs);
+        return this.paginate(this.parentJobs);
       },
       displayedPages(){
         var from=this.page-1 // from the first
@@ -260,6 +301,32 @@
   }
 </script>
 <style scoped>
+
+  .table td,.table th{
+    max-width: 0;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+  .table,table tr {
+    width: 100%!important;
+  }
+  table thead th.action{
+    width:3em!important;
+    max-width:3em!important;
+  }
+  table thead th.id{
+    width:6em!important;
+    max-width:6em!important;
+  }
+  table thead th.jobtype{
+    width:6em!important;
+    max-width:6em!important;
+  }
+  table thead th.status{
+    width:6em!important;
+    max-width:6em!important;
+  }
   .cursor-progress{
     cursor:progress;
   }