+++ server/config/ansible.config.js
@@ -1,6 +1,6 @@
 const resolve = require("path").resolve
 var ansible_config = {
-  path: process.env.ANSIBLE_PATH || (__dirname + '/../persistent/playbooks')
+  path: process.env.ANSIBLE_PATH || resolve(__dirname + '/../persistent/playbooks')
 };
 
 module.exports = ansible_config;