+++ server/src/configure.js
@@ -40,7 +40,7 @@ module.exports = app => {
   const jobRoutes = require('./routes/job.routes')
   const ansibleRoutes = require('./routes/ansible.routes')
   const gitRoutes = require('./routes/git.routes')
-  const multiRoutes = require('./routes/multi.routes')
+  const multistepRoutes = require('./routes/multistep.routes')
   const queryRoutes = require('./routes/query.routes')
   const expressionRoutes = require('./routes/expression.routes')
   const userRoutes = require('./routes/user.routes')
@@ -54,6 +54,8 @@ module.exports = app => {
   const versionRoutes = require('./routes/version.routes')
   const profileRoutes = require('./routes/profile.routes')
   const sshRoutes = require('./routes/ssh.routes')
+  const logRoutes = require('./routes/log.routes')
+  const repoRoutes = require('./routes/repo.routes')
 
   // using json web tokens as middleware
   const authobj = passport.authenticate('jwt', { session: false })
@@ -86,7 +88,7 @@ module.exports = app => {
   app.use('/api/v1/awx',cors(), authobj, awxRoutes) // extra middleware in the routes
   app.use('/api/v1/ansible',cors(), authobj, ansibleRoutes)
   app.use('/api/v1/git',cors(), authobj, gitRoutes)
-  app.use('/api/v1/multi',cors(), authobj, multiRoutes)
+  app.use('/api/v1/multistep',cors(), authobj, multistepRoutes)
 
   // api routes for admin management
   app.use('/api/v1/job',cors(), authobj, jobRoutes)
@@ -95,6 +97,8 @@ module.exports = app => {
   app.use('/api/v1/ldap',cors(), authobj, checkAdminMiddleware, ldapRoutes)
   app.use('/api/v1/credential',cors(), authobj, checkAdminMiddleware, credentialRoutes)
   app.use('/api/v1/sshkey',cors(), authobj, checkAdminMiddleware, sshRoutes)
+  app.use('/api/v1/log',cors(), authobj, checkAdminMiddleware, logRoutes)
+  app.use('/api/v1/repo',cors(), authobj, checkAdminMiddleware, repoRoutes)
 
   // routes for form config (extra middleware in the routes itself)
   app.use('/api/v1/config',cors(), authobj, configRoutes)