+++ server/src/models/git.model.js
@@ -4,6 +4,7 @@ const path=require("path")
 const Exec=require("../lib/exec.js")
 const Job=require("./job.model")
 const YAML=require("yaml")
+var config=require('../../config/app.config')
 const fs=require("fs")
 
 var Git=function(){
@@ -27,21 +28,23 @@ Git.push = function (form,repo,extraVars,user, result,success,failed) {
       // rest is in background - launch save + git commit/push
       try{
         // save the extravars as file - we do this in sync, should be fast
+        Exec.printCommand(`TASK [Writing YAML to local repo] ${'*'.repeat(72-26)}`,"stdout",jobid,++counter)
         var yaml = YAML.stringify(JSON.parse(extraVars))
-        fs.writeFileSync(path.join(repo.dir,repo.file),yaml)
+        fs.writeFileSync(path.join(config.repoDir,repo.dir,repo.file),yaml)
         // log the save
-        Exec.printCommand(`Extravars Yaml file created : ${repo.file}`,"stdout",jobid,counter,function(){
-          // start commit
-          var command = `git commit --allow-empty -am "update from ansibleforms" && ${repo.push}`
-          var directory = repo.dir
-          Exec.executeCommand({directory:directory,command:command,description:"Pushing to git",task:"Git push"},jobid,counter,(jobid,counter)=>{
-            if(success)success()
-          },(jobid,counter)=>{
-            if(failed)failed("Job failed")
-          },(jobid,counter)=>{
-            if(failed)failed("Job was aborted")
-          })
+        Exec.printCommand(`ok: [Extravars Yaml file created : ${repo.file}]`,"stdout",jobid,++counter)
+        Exec.printCommand(`TASK [Committing changes] ${'*'.repeat(72-18)}`,"stdout",jobid,++counter)
+        // start commit
+        var command = `git commit --allow-empty -am "update from ansibleforms" && ${repo.push}`
+        var directory = path.join(config.repoDir,repo.dir)
+        Exec.executeCommand({directory:directory,command:command,description:"Pushing to git",task:"Git push"},jobid,counter,(jobid,counter)=>{
+          if(success)success()
+        },(jobid,counter)=>{
+          if(failed)failed("Job failed")
+        },(jobid,counter)=>{
+          if(failed)failed("Job was aborted")
         })
+
       }catch(e){
         logger.error(e)
         Exec.endCommand(jobid,counter,"stderr","failed",e,null,(jobid,counter)=>{
@@ -57,7 +60,7 @@ Git.push = function (form,repo,extraVars,user, result,success,failed) {
 Git.pull = function (repo, result) {
 
     var command = repo.pull
-    var directory = repo.dir
+    var directory = path.join(config.repoDir,repo.dir)
     Exec.executeSilentCommand({directory:directory,command:command,description:"Pulling from git"},result)
 
 };