+++ server/src/routes/multi.routes.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const router = express.Router()
-const multiController =   require('../controllers/multi.controller');
-
-// push to git
-router.post('/launch', multiController.launch);
-
-module.exports = router