+++ server/src/routes/multistep.routes.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const router = express.Router()
+const multistepController =   require('../controllers/multistep.controller');
+
+// push to git
+router.post('/launch', multistepController.launch);
+
+module.exports = router